slqlite3: debe y haber se cambiaron a INTEGER | modelo: todos los int son int64_t

+++ include/Asiento.h
@@ -1,18 +1,18 @@
 #ifndef ASIENTO_H
 #define ASIENTO_H
 #include <string>
-
+#include <cstdint>
 class Asiento {
 private:
     std::string fecha;
     std::string comentarios;
-    int id;
+    int64_t id;
 public:
-    Asiento(int id, std::string fecha,std::string comentarios);
-    int getId();
+    Asiento(int64_t id, std::string fecha,std::string comentarios);
+    int64_t getId();
     std::string getFecha();
     std::string getComentarios();
-    void setId(int id);
+    void setId(int64_t id);
 };
 
 #endif
