+++ include/AsientoDAO.h
@@ -2,6 +2,7 @@
 #define ASIENTODAO_H
 #include "sqlite3.h"
 #include <vector>
+#include <cstdint>
 
 class Asiento;
 
@@ -10,8 +11,8 @@ class AsientoDAO {
 
 public:
     static void guardar(sqlite3* db, Asiento* asiento);
-    static void eliminar(sqlite3* db, int id);
-    static void actualizar(sqlite3* db,int id);
+    static void eliminar(sqlite3* db, int64_t id);
+    static void actualizar(sqlite3* db,int64_t id);
     static void cargarAsientos(sqlite3* db, std::vector<Asiento*> &Asientos);
     static std::vector<Asiento*> cargarTodas(sqlite3* db);
 };
