+++ include/Cuenta.h
@@ -1,23 +1,23 @@
 #ifndef CUENTA_H
 #define CUENTA_H
 #include <string>
-
+#include <cstdint>
 class Cuenta {
 private:
     std::string nombre;
     std::string descripcion;
     std::string rubro;
-    int numero; // este es el nº elegido por el usuario NO ID
-    int id;
+    int64_t numero; // este es el nº elegido por el usuario NO ID
+    int64_t id;
 public:
-    Cuenta(int id,std::string nombre, std::string descripcion,std::string rubro, int numero);
+    Cuenta(int64_t id,std::string nombre, std::string descripcion,std::string rubro, int64_t numero);
     std::string getNombre();
     std::string getDescripcion();
     std::string getRubro();
-    int getId();
-    int getNumero();
-    void setNumero(int numero);
-    void setId(int id);
+    int64_t getId();
+    int64_t getNumero();
+    void setNumero(int64_t numero);
+    void setId(int64_t id);
 };
 
 #endif
