+++ src/Asiento.cpp
@@ -1,12 +1,12 @@
 #include "Asiento.h"
 
-Asiento::Asiento(int id, std::string fecha, std::string comentarios){
+Asiento::Asiento(int64_t id, std::string fecha, std::string comentarios){
     this->id=id;
     this->fecha=fecha;
     this->comentarios=comentarios;
 }
 
-int Asiento::getId(){
+int64_t Asiento::getId(){
     return id;
 }
 
@@ -18,6 +18,6 @@ std::string Asiento::getComentarios(){
     return comentarios;
 }
 
-void Asiento::setId(int id){
+void Asiento::setId(int64_t id){
     this->id=id;
 }
