+++ src/Registro.cpp
@@ -1,6 +1,6 @@
 #include "Registro.h"
 
-Registro::Registro(int id, int cuentaId, int asientoId, std::string notas, float debe, float haber){
+Registro::Registro(int64_t id, int64_t cuentaId, int64_t asientoId, std::string notas, int64_t debe, int64_t haber){
         this->notas=notas;
         this->cuentaId=cuentaId;
         this->id=id;
@@ -13,26 +13,26 @@ std::string Registro::getNotas(){
     return notas;
 }
 
-int Registro::getId(){
+int64_t Registro::getId(){
     return id;
 }
 
-int Registro::getCuentaId(){
+int64_t Registro::getCuentaId(){
     return cuentaId;
 }
 
-int Registro::getAsientoId(){
+int64_t Registro::getAsientoId(){
     return asientoId;
 }
 
-void Registro::setId(int id){
+void Registro::setId(int64_t id){
     this->id=id;
 }
 
-float Registro::getDebe(){
+int64_t Registro::getDebe(){
     return debe;
 }
 
-float Registro::getHaber(){
+int64_t Registro::getHaber(){
     return haber;
 }
